custom-elements/instantiating-custom-elements/extensions-to-document-interface/create-element-test-001.html

<!DOCTYPE html>
<html>
<head>
<title>If typeExtension is not present then type is a local name</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If typeExtension is not present then type is a local name">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#extensions-to-document-interface-to-instantiate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    var name = 'x-a-1';
    
    doc.register(name);

    var customElement = doc.createElement(name);

    assert_equals(customElement.constructor.prototype.constructor.name, name, 
            'Wrong custom element type');    

}, 'If typeExtension is not present in createElement call then type is a local name');


test(function() {
    var doc = newHTMLDocument();
    
    var name = 'x-a-2';
    
    doc.register(name);

    var customElement = doc.createElementNS(HTML_NAMESPACE, name);

    assert_equals(customElement.constructor.prototype.constructor.name, name, 
            'Wrong custom element type');

}, 'If typeExtension is not present in createElementNS call then type is a local name');


</script>
</body>
</html>
